+++ R/helpers_ggcoefstats.R
@@ -482,8 +482,10 @@ tfz_labeller <- function(tidy_df,
 #'   (corresponding estimates of coefficients or other quantities of interest)
 #'   and `std.error` (the standard error of the regression term).
 #' @param output  Character describing the desired output. If `"subtitle"`, a
-#'   formatted subtitle will be returned. The other option is to return `"tidy"`
-#'   data frame with coefficients.
+#'   formatted subtitle with summary effect and statistical details will be
+#'   returned, and if `"caption"`, expression containing details from model
+#'   summary will be returned. The other option is to return `"tidy"` data frame
+#'   with coefficients or `"glance"` dataframe with model summaries.
 #' @inheritParams ggbetweenstats
 #' @param ... Additional arguments (ignored).
 #'
@@ -492,29 +494,33 @@ tfz_labeller <- function(tidy_df,
 #' @examples
 #' # let's create a dataframe
 #' df_results <-
-#'   structure(list(estimate = c(
-#'     0.382047603321706, 0.780783111514665,
-#'     0.425607573765058, 0.558365541235078, 0.956473848429961
-#'   ), std.error = c(
-#'     0.0465576338644502,
-#'     0.0330218199731529, 0.0362834986178494, 0.0480571500648261, 0.062215818388157
-#'   ), t.value = c(
-#'     8.20590677855356, 23.6444603038067, 11.7300588415607,
-#'     11.6187818146078, 15.3734833553524
-#'   ), conf.low = c(
-#'     0.290515146096969,
-#'     0.715841986960399, 0.354354575031406, 0.46379116008131, 0.827446138277154
-#'   ), conf.high = c(
-#'     0.473580060546444, 0.845724236068931, 0.496860572498711,
-#'     0.652939922388847, 1.08550155858277
-#'   ), p.value.x = c(
-#'     3.28679518728519e-15,
-#'     4.04778497135963e-75, 7.59757330804449e-29, 5.45155840151592e-26,
-#'     2.99171217913312e-13
-#'   ), df.residual = c(
-#'     394L, 358L, 622L, 298L,
-#'     22L
-#'   )), row.names = c(NA, -5L), class = c("tbl_df", "tbl", "data.frame"))
+#'   structure(
+#'     .Data = list(estimate = c(
+#'       0.382047603321706, 0.780783111514665,
+#'       0.425607573765058, 0.558365541235078, 0.956473848429961
+#'     ), std.error = c(
+#'       0.0465576338644502,
+#'       0.0330218199731529, 0.0362834986178494, 0.0480571500648261, 0.062215818388157
+#'     ), t.value = c(
+#'       8.20590677855356, 23.6444603038067, 11.7300588415607,
+#'       11.6187818146078, 15.3734833553524
+#'     ), conf.low = c(
+#'       0.290515146096969,
+#'       0.715841986960399, 0.354354575031406, 0.46379116008131, 0.827446138277154
+#'     ), conf.high = c(
+#'       0.473580060546444, 0.845724236068931, 0.496860572498711,
+#'       0.652939922388847, 1.08550155858277
+#'     ), p.value.x = c(
+#'       3.28679518728519e-15,
+#'       4.04778497135963e-75, 7.59757330804449e-29, 5.45155840151592e-26,
+#'       2.99171217913312e-13
+#'     ), df.residual = c(
+#'       394L, 358L, 622L, 298L,
+#'       22L
+#'     )),
+#'     row.names = c(NA, -5L),
+#'     class = c("tbl_df", "tbl", "data.frame")
+#'   )
 #'
 #' # making subtitle
 #' ggstatsplot::subtitle_meta_ggcoefstats(
@@ -529,6 +535,21 @@ tfz_labeller <- function(tidy_df,
 #'   messages = FALSE,
 #'   output = "tidy"
 #' )
+#'
+#' # making caption
+#' ggstatsplot::subtitle_meta_ggcoefstats(
+#'   data = df_results,
+#'   k = 2,
+#'   messages = FALSE,
+#'   output = "caption"
+#' )
+#'
+#' # getting dataframe with model summary
+#' ggstatsplot::subtitle_meta_ggcoefstats(
+#'   data = df_results,
+#'   messages = FALSE,
+#'   output = "glance"
+#' )
 #' @export
 
 # function body
@@ -536,7 +557,10 @@ subtitle_meta_ggcoefstats <- function(data,
                                       k = 2,
                                       messages = TRUE,
                                       output = "subtitle",
+                                      caption = NULL,
                                       ...) {
+
+  #----------------------- input checking ------------------------------------
 
   # check if the two columns needed are present
   if (sum(c("estimate", "std.error") %in% names(data)) != 2) {
@@ -550,6 +574,8 @@ subtitle_meta_ggcoefstats <- function(data,
     call. = FALSE
     )
   }
+
+  #----------------------- meta-analysis ------------------------------------
 
   # object from meta-analysis
   meta_res <- metafor::rma(
@@ -572,8 +598,10 @@ subtitle_meta_ggcoefstats <- function(data,
     print(summary(meta_res))
   }
 
+  #----------------------- tidy output and subtitle ---------------------------
+
   # create a dataframe with coeffcients
-  df_coef <- coef(summary(meta_res)) %>%
+  df_tidy <- coef(summary(meta_res)) %>%
     tibble::as_tibble(x = .) %>%
     dplyr::rename(
       .data = .,
@@ -598,7 +626,7 @@ subtitle_meta_ggcoefstats <- function(data,
     base::substitute(
       expr =
         paste(
-          "Meta-analytic effect: ",
+          "Summary effect: ",
           beta,
           " = ",
           estimate,
@@ -621,20 +649,79 @@ subtitle_meta_ggcoefstats <- function(data,
           pvalue
         ),
       env = base::list(
-        estimate = specify_decimal_p(x = df_coef$estimate, k = k),
-        LL = specify_decimal_p(x = df_coef$conf.low, k = k),
-        UL = specify_decimal_p(x = df_coef$conf.high, k = k),
-        zvalue = specify_decimal_p(x = df_coef$z.value, k = k),
-        se = specify_decimal_p(x = df_coef$std.error, k = k),
-        pvalue = specify_decimal_p(x = df_coef$p.value, k = k, p.value = TRUE)
+        estimate = specify_decimal_p(x = df_tidy$estimate, k = k),
+        LL = specify_decimal_p(x = df_tidy$conf.low, k = k),
+        UL = specify_decimal_p(x = df_tidy$conf.high, k = k),
+        zvalue = specify_decimal_p(x = df_tidy$z.value, k = k),
+        se = specify_decimal_p(x = df_tidy$std.error, k = k),
+        pvalue = specify_decimal_p(x = df_tidy$p.value, k = k, p.value = TRUE)
       )
     )
 
+  #----------------------- input checking ------------------------------------
+
+  df_glance <- with(
+    data = meta_res,
+    expr = tibble::tibble(
+      tau2 = tau2,
+      se.tau2 = se.tau2,
+      k = k,
+      p = p,
+      m = m,
+      QE = QE,
+      QEp = QEp,
+      QM = QM,
+      QMp = QMp,
+      I2 = I2,
+      H2 = H2,
+      int.only = int.only
+    )
+  )
+
+  # preparing the subtitle
+  caption <-
+    base::substitute(
+      atop(displaystyle(top.text),
+        expr =
+          paste(
+            "Heterogeneity: ",
+            italic("Q"),
+            "(",
+            df,
+            ") = ",
+            Q,
+            ", ",
+            italic("p "),
+            pvalue,
+            ", ",
+            tau["REML"]^2,
+            " = ",
+            tau2,
+            ", ",
+            "I"^2,
+            " = ",
+            I2
+          )
+      ),
+      env = base::list(
+        top.text = caption,
+        Q = specify_decimal_p(x = df_glance$QE, k = 0L),
+        df = specify_decimal_p(x = (df_glance$k - 1), k = 0L),
+        pvalue = specify_decimal_p(x = df_glance$QEp, k = k, p.value = TRUE),
+        tau2 = specify_decimal_p(x = df_glance$tau2, k = k),
+        I2 = paste(specify_decimal_p(x = df_glance$I2, k = 2L), "%", sep = "")
+      )
+    )
+
+  #---------------------------- output ---------------------------------------
+
   if (output == "subtitle") {
-    # return the subtitle
     return(subtitle)
   } else if (output == "tidy") {
-    # tidy data frame with coefficients
-    return(df_coef)
+    return(df_tidy)
+  } else if (output == "caption") {
+    return(caption)
+  } else if (output == "glance") {
+    return(df_glance)
   }
 }