+++ tests/testthat/test_subtitle_meta.R
@@ -5,6 +5,7 @@ context("subtitle_meta_ggcoefstats")
 testthat::test_that(
   desc = "subtitle_meta_ggcoefstats works",
   code = {
+    testthat::skip_on_cran()
 
     # dataframe
     df <- tibble::tribble(
@@ -17,6 +18,7 @@ testthat::test_that(
     )
 
     # subtitle output
+    set.seed(123)
     using_function1 <- ggstatsplot::subtitle_meta_ggcoefstats(
       data = df,
       k = 4,
@@ -24,12 +26,22 @@ testthat::test_that(
       output = "subtitle"
     )
 
-    # expected output
+    # caption output
+    set.seed(123)
+    using_function2 <- ggstatsplot::subtitle_meta_ggcoefstats(
+      data = df,
+      k = 2,
+      caption = "this is caption",
+      messages = FALSE,
+      output = "caption"
+    )
+
+    # expected subtitle output
     set.seed(123)
     results1 <-
       ggplot2::expr(
         paste(
-          "Meta-analytic effect: ",
+          "Summary effect: ",
           beta,
           " = ",
           "0.4377",
@@ -53,23 +65,64 @@ testthat::test_that(
         )
       )
 
+    # expected subtitle output
+    set.seed(123)
+    results2 <-
+      ggplot2::expr(atop(
+        displaystyle("this is caption"),
+        expr = paste(
+          "Heterogeneity: ",
+          italic("Q"),
+          "(",
+          "4",
+          ") = ",
+          "6083",
+          ", ",
+          italic("p "),
+          "< 0.001",
+          ", ",
+          tau["REML"]^2,
+          " = ",
+          "0.19",
+          ", ",
+          "I"^2,
+          " = ",
+          "99.90%"
+        )
+      ))
+
     # testing overall call
     testthat::expect_identical(using_function1, results1)
+    testthat::expect_identical(using_function2, results2)
 
     # tidy dataframe
     set.seed(123)
     tidy_df <- ggstatsplot::subtitle_meta_ggcoefstats(
       data = df,
-      k = 4,
       messages = FALSE,
       output = "tidy"
     )
 
-    # checking if the output is expected
+    glance_df <- ggstatsplot::subtitle_meta_ggcoefstats(
+      data = df,
+      messages = FALSE,
+      output = "glance"
+    )
+
+    # checking if the tidy output is expected
     testthat::expect_equal(dim(tidy_df), c(1L, 7L))
     testthat::expect_equal(tidy_df$estimate, 0.4376927, tolerance = 0.0001)
     testthat::expect_equal(tidy_df$conf.low, 0.04231262, tolerance = 0.0001)
     testthat::expect_equal(tidy_df$conf.high, 0.8330728, tolerance = 0.0001)
+
+    # checking if the glance output is expected
+    testthat::expect_equal(dim(glance_df), c(1L, 12L))
+    testthat::expect_equal(glance_df$tau2, 0.193525, tolerance = 0.0001)
+    testthat::expect_equal(glance_df$se.tau2, 0.1437787, tolerance = 0.0001)
+    testthat::expect_equal(glance_df$QE, 6082.595, tolerance = 0.01)
+    testthat::expect_equal(glance_df$k, 5L)
+    testthat::expect_equal(glance_df$p, 1L)
+    testthat::expect_equal(glance_df$m, 1L)
   }
 )
 
@@ -79,6 +132,7 @@ testthat::test_that(
 testthat::test_that(
   desc = "checking meta-analysis results object",
   code = {
+    testthat::skip_on_cran()
 
     # dataframe
     df <- tibble::tribble(